Time: 363 ms (45.98%), Space: 9.1 MB (65.53%) - LeetHub

+++ 2318-number-of-distinct-roll-sequences/2318-number-of-distinct-roll-sequences.cpp
@@ -7,7 +7,7 @@ public:
             return 1;
         }
         if(memo[n][prev][prevprev]==-1){
-            long long int res=0;
+            int res=0;
             for(int i=1;i<=6;i++){
                 if(i!=prev && i!=prevprev && (prev==0 || __gcd(prev,i)==1)){
                     res=(res%mod + fun(n-1,i,prev)%mod)%mod;
